pyproject.toml
[tool.poetry]
name = "oximeter-telegram-bot"
version = "0.1.0"
description = ""
authors = ["LaBatata101 <labatata101@linuxmail.org>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = {version = "^20*", allow-prereleases =  true}
aiohttp = {extras = ["speedups"], version = "^3.8.1"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
bot = "oximeter.bot:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
